+++ crates/ash-parser/tests/task_1758_macro_lowering_boundaries.rs
@@ -1,5 +1,7 @@
 use ash_parser::lower::{expand_and_lower_surface_module, lower_expanded_surface_module};
-use ash_parser::surface::{ExpandedSurfaceModule, Expr, MacroDelimiter, MacroInvocation};
+use ash_parser::surface::{
+    ExpandedSurfaceModule, Expr, MacroDelimiter, MacroInvocation, MacroInvocationBody,
+};
 use ash_parser::token::Span;
 
 #[test]
@@ -46,6 +48,7 @@ fn direct_expanded_gate_rejects_raw_macro_carriers_if_caller_constructs_invalid_
             name: "inc".into(),
             delimiter: MacroDelimiter::Paren,
             raw_body: "n".into(),
+            body: MacroInvocationBody::ExprArgs(Vec::new()),
             token_trees: Vec::new(),
             args: Some(Vec::new()),
             span: Span::new(0, 1, 1, 1),